% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{.make_feature_names}
\alias{.make_feature_names}
\title{Make feature names}
\usage{
.make_feature_names(features_list)
}
\arguments{
\item{features_list}{list with feature names of functions that
generate them}
}
\value{
character vector with feature names
}
\description{
Make feature names
}
