% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encode_ordinal}
\alias{encode_ordinal}
\title{Encode ordinal}
\usage{
encode_ordinal(x, order = sort(unique(x)))
}
\arguments{
\item{x}{vector to encode}

\item{order}{vector with x elements in order of encoding}
}
\value{
numeric vector with encoded values
}
\description{
to integer values
}
