% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{extract_features}
\alias{extract_features}
\alias{extract_features.neuronlist}
\title{Extract features from neuronlist}
\usage{
extract_features(
  neurons_list,
  features_list = NULL,
  y = NULL,
  to_numeric = FALSE,
  normalise = c("none", "zscore", "scale")
)

\method{extract_features}{neuronlist}(
  neurons_list,
  features_list = NULL,
  y = NULL,
  to_numeric = FALSE,
  normalise = c("none", "zscore", "scale"),
  ...
)
}
\arguments{
\item{neurons_list}{neuronlist of neurons (nat)}

\item{features_list}{list with feature names of functions that
generate them}

\item{y}{target class (can be character of function), will be added as column
"Y" if name is NULL.}

\item{to_numeric}{flag that says whether all factor/ character features are
meant to be transformed to numeric (default: F)}

\item{normalise}{what normalization to use}

\item{...}{extra arguments}
}
\value{
data frame with features
}
\description{
extract features from a neuronlist
}
\examples{
features_list <- list( "upstream", "downstream", "soma")
kc_train_features <- extract_features(
  kc_train,
  features_list = features_list,
  y = "type"
)
}
