% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{get_cable_length}
\alias{get_cable_length}
\title{Get cable length}
\usage{
get_cable_length(n)
}
\arguments{
\item{n}{(nat) neuron}
}
\value{
numeric with total cable length
}
\description{
Get cable length
}
\seealso{
Other neuron_scalars: 
\code{\link{get_n_branchpoints}()},
\code{\link{get_radius}()},
\code{\link{get_soma_size}()},
\code{\link{get_spine_length}()},
\code{\link{get_spine_radius}()}
}
\concept{neuron_scalars}
