% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{get_n_branchpoints}
\alias{get_n_branchpoints}
\title{Get number of branchpoints}
\usage{
get_n_branchpoints(n)
}
\arguments{
\item{n}{(nat) neuron}
}
\value{
numeric with soma size
}
\description{
Get number of branchpoints
}
\seealso{
Other neuron_scalars: 
\code{\link{get_cable_length}()},
\code{\link{get_radius}()},
\code{\link{get_soma_size}()},
\code{\link{get_spine_length}()},
\code{\link{get_spine_radius}()}
}
\concept{neuron_scalars}
