% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{get_soma_size}
\alias{get_soma_size}
\title{Get soma size}
\usage{
get_soma_size(n)
}
\arguments{
\item{n}{(nat) neuron}
}
\value{
numeric with soma size
}
\description{
Get soma size
}
\seealso{
Other neuron_scalars: 
\code{\link{get_cable_length}()},
\code{\link{get_n_branchpoints}()},
\code{\link{get_radius}()},
\code{\link{get_spine_length}()},
\code{\link{get_spine_radius}()}
}
\concept{neuron_scalars}
