% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{get_spine_length}
\alias{get_spine_length}
\title{Get a neuron spine length}
\usage{
get_spine_length(n)
}
\arguments{
\item{n}{(nat) neuron}
}
\value{
numeric with spine length
}
\description{
Get a neuron spine length
}
\seealso{
Other neuron_scalars: 
\code{\link{get_cable_length}()},
\code{\link{get_n_branchpoints}()},
\code{\link{get_radius}()},
\code{\link{get_soma_size}()},
\code{\link{get_spine_radius}()}
}
\concept{neuron_scalars}
