% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{get_spine_radius}
\alias{get_spine_radius}
\title{Get average spine radius}
\usage{
get_spine_radius(n)
}
\arguments{
\item{n}{(nat) neuron}
}
\value{
numeric with soma size
}
\description{
Get average spine radius
}
\seealso{
Other neuron_scalars: 
\code{\link{get_cable_length}()},
\code{\link{get_n_branchpoints}()},
\code{\link{get_radius}()},
\code{\link{get_soma_size}()},
\code{\link{get_spine_length}()}
}
\concept{neuron_scalars}
