% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kc_train}
\alias{kc_train}
\alias{kc_test}
\title{Subset of Hemibrain Kenyon Cell neurons}
\format{
An object of class \code{neuronlist} (inherits from \code{list}) of length 300.

An object of class \code{neuronlist} (inherits from \code{list}) of length 75.
}
\usage{
kc_train

kc_test
}
\description{
Separated in balances training (300 examples) and test (75 examples) splits.
}
\details{
Note that these positions are raw (8nm voxel) locations.
}
\examples{

library(nat)
library(natML)
plot3d(kc_train)
}
\references{
Scheffer, L.K., Xu, C.S., Januszewski, M., Lu, Z., Takemura, S.Y.,
Hayworth, K.J., Huang, G.B., Shinomiya, K., Maitlin-Shepard, J., Berg, S.
and Clements, J., 2020. A connectome and analysis of the adult Drosophila
central brain. Elife, 9, p.e57443.
}
\keyword{datasets}
