% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale}
\alias{scale}
\title{Scale vector from 0 to 1}
\usage{
scale(vec, na_rm = TRUE)
}
\arguments{
\item{vec}{vector with numbers}

\item{na_rm}{flag stating whether to ignore NA values or not}
}
\value{
vector with scaled values
}
\description{
Note that vector with all the same values will be filled with 0.5.
}
\seealso{
Other scaling: 
\code{\link{zscore}()}
}
\concept{scaling}
