% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{split_test_train}
\alias{split_test_train}
\title{Split into test train sets}
\usage{
split_test_train(yclass, split_prop)
}
\arguments{
\item{yclass}{any vector with Y labels}

\item{split_prop}{proportion of test samples}
}
\value{
indices of train test splits
}
\description{
It makes sure that there is at least one representative of each class
in the test test
}
\examples{
y <- c(1,1,1,1,2,2,2,2,2,2,2,3,3)
split_test_train(y,0.2)
}
