% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zscore}
\alias{zscore}
\title{Z score vector}
\usage{
zscore(vec, na_rm = TRUE)
}
\arguments{
\item{vec}{vector with numbers}

\item{na_rm}{flag stating whether to ignore NA values or not}
}
\value{
vector with z-scored values
}
\description{
Note that vector with all the same values will be filled with 0.
}
\seealso{
Other scaling: 
\code{\link{scale}()}
}
\concept{scaling}
