% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{.get_feature}
\alias{.get_feature}
\title{Get feature}
\usage{
.get_feature(neurons_list, feature, as_factor = FALSE, n_jobs = 1)
}
\arguments{
\item{neurons_list}{neuronlist of neurons}

\item{feature}{callable or character with feature}

\item{as_factor}{flag saying whether feature should be transformed to factor}

\item{n_jobs}{number of cores used to call a function feature extraction}
}
\value{
vector with feature values
}
\description{
from a neuron list.
}
